+++ src/Raven.Server/Documents/Expiration/ExpirationStorage.cs
@@ -80,20 +80,20 @@ namespace Raven.Server.Documents.Expiration
         }
 
         public Dictionary<Slice, List<(Slice LowerId, string Id)>> GetExpiredDocuments(DocumentsOperationContext context,
-            DateTime currentTime, bool applyToExistingDocuments, int take, out int count, out Stopwatch duration, CancellationToken cancellationToken)
+            DateTime currentTime, bool applyToExistingDocuments, long take, out long count, out Stopwatch duration, CancellationToken cancellationToken)
         {
             return GetDocuments(context, currentTime, applyToExistingDocuments, DocumentsByExpiration, Constants.Documents.Metadata.Expires, take, out count, out duration, cancellationToken);
         }
 
         public Dictionary<Slice, List<(Slice LowerId, string Id)>> GetDocumentsToRefresh(DocumentsOperationContext context,
-            DateTime currentTime, bool applyToExistingDocuments, int take, out int count, out Stopwatch duration, CancellationToken cancellationToken)
+            DateTime currentTime, bool applyToExistingDocuments, long take, out long count, out Stopwatch duration, CancellationToken cancellationToken)
 
         {
             return GetDocuments(context, currentTime, applyToExistingDocuments, DocumentsByRefresh, Constants.Documents.Metadata.Refresh, take, out count, out duration, cancellationToken);
         }
 
         private Dictionary<Slice, List<(Slice LowerId, string Id)>> GetDocuments(DocumentsOperationContext context,
-            DateTime currentTime, bool applyToExistingDocuments, string treeName, string metadataPropertyToCheck, int take, out int count, out Stopwatch duration, CancellationToken cancellationToken)
+            DateTime currentTime, bool applyToExistingDocuments, string treeName, string metadataPropertyToCheck, long take, out long count, out Stopwatch duration, CancellationToken cancellationToken)
         {
             count = 0;
             var currentTicks = currentTime.Ticks;
